reference/experiments/sportsgameodds/requirements.txt
# SportsGameOdds WebSocket Bot dependencies
pysher>=1.0.8
requests>=2.28.0
python-telegram-bot>=20.0
python-dotenv>=1.0.0
